Response is an array of code/message, not strings
Add a new ResponseInfo struct containing the Code and Message fields,
have the Response struct return a slice of ResponseInfo instead of a
slice of strings.

It seems the response info can also have a field "type" but I've yet to
see it return anything other than null, so omitting that for now.